+++ src/eyed3/info.py
@@ -21,11 +21,11 @@ import os
 NAME            = "eyeD3"
 VERSION         = "0.7.5"
 VERSION_TUPLE   = tuple((int(v) for v in VERSION.split('.')))
-RELEASE         = "beta"
+RELEASE         = "final"
 AUTHOR          = "Travis Shirk"
-URL             = "http://eyeD3.nicfit.net"
+URL             = "http://eyeD3.nicfit.net/"
 VERSION_MSG     = """
-eyeD3 %s-%s (C) Copyright 2002-2013 %s
+eyeD3 %s-%s (C) Copyright 2002-2014 %s
 This program comes with ABSOLUTELY NO WARRANTY! See COPYING for details.
 Run with --help/-h for usage information or read the docs at
 %s